C++ warning fixes

+++ libs/drape_frontend/visual_params.cpp
@@ -180,7 +180,7 @@ int GetTileScaleBase(m2::RectD const & r)
 {
   double const sz = std::max(r.SizeX(), r.SizeY());
   ASSERT_GREATER(sz, 0., ("Rect should not be a point:", r));
-  return std::max(1l, std::lround(std::log2(mercator::Bounds::kRangeX / sz)));
+  return std::max(1, static_cast<int>(std::lround(std::log2(mercator::Bounds::kRangeX / sz))));
 }
 
 double GetTileScaleBase(double drawScale)
